Revert "Optimised all the imports and reformatted the arranged imports for better code readibility"
This reverts commit e599151f98bdbe05690853673451f6d73729d3bd.

+++ AnkiDroid/src/test/java/com/ichi2/libanki/DecksTest.java
@@ -95,7 +95,7 @@ public class DecksTest extends RobolectricTest {
         assertNotEquals(parentId, 0);
         assertEquals(2, decks.allSortedNames().size());
         // should get the same id
-        assertEquals(parentId, addDeck("new deck"));
+        assertEquals(parentId, (long) addDeck("new deck"));
         // we start with the default col selected
         assertEquals(1, decks.selected());
         assertEqualsArrayList(new Long[] {1L}, decks.active());
